requirements.txt
apache-airflow==2.7.0
yfinance==0.2.28